requirements.txt
Django==3.1
django-cors-headers
django-filter
django-storages
djangorestframework
djangorestframework-csv
djangorestframework-jwt
djangorestframework-simplejwt
djoser==2.0.3
dnspython==1.16.0
httplib2==0.15.0
requests==2.22.0
django-redis

celery==4.4.1
redis==3.4.1
flower==0.9.3
